Reorg Number Instances
