--- kategory/src/main/kotlin/kategory/instances/NumberSemigroup.kt
@@ -1,6 +1,0 @@
-package kategory
-
-class NumberSemigroup<A : Number>(val f: (A, A) -> A) : Semigroup<A> {
-    override fun combine(a: A, b: A): A =
-            f(a, b)
-}
